+++ src/batou/lib/tests/test_file.py
@@ -839,7 +839,9 @@ def test_mode_verifies_for_nonexistent_file(root):
         mode.verify()
 
 
-def test_mode_ensures_mode_for_files(root):
+@pytest.mark.parametrize('input,expected', [
+    (0o777, 0o777),])
+def test_mode_ensures_mode_for_files(root, input, expected):
     path = "path"
     open("work/mycomponent/" + path, "w").close()
     mode = Mode(path, mode=0o000)