+++ src/batou/tests/fixture/basic_service/components/zeo/component.py
@@ -3,6 +3,6 @@ from batou.component import Component, Attribute
 
 class ZEO(Component):
 
-    port = Attribute(int, "9001")
+    port = Attribute(int, default_conf_string="9001")
 
     features = ["test", "test2"]